+++ src/tests/models/test_inventory.py
